prototype_for_testing/requirements.txt
absl-py==2.3.1
attrs==25.4.0
cffi==2.0.0
coloredlogs==15.0.1
contourpy==1.3.3
cycler==0.12.1
flatbuffers==25.9.23
fonttools==4.60.1
humanfriendly==10.0
jax==0.6.1
jaxlib==0.6.1
kiwisolver==1.4.9
matplotlib==3.10.7
mediapipe==0.10.21
ml_dtypes==0.5.3
mpmath==1.3.0
numpy==1.26.4
onnx==1.16.1
onnxruntime==1.19.2
opencv-contrib-python==4.11.0.86
opencv-python==4.10.0.84
openvino==2025.3.0
openvino-telemetry==2025.2.0
opt_einsum==3.4.0
packaging==25.0
pillow==11.3.0
protobuf==4.25.3
pycparser==2.23
pyparsing==3.2.5
python-dateutil==2.9.0.post0
scipy==1.11.4
sentencepiece==0.2.1
six==1.17.0
sounddevice==0.5.2
sympy==1.14.0
